requirements.txt
flask>=3.0.0
line-bot-sdk>=3.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
apscheduler>=3.10.0
